refactor: доработать контроллеры для Category
