+++ ewm-service/src/main/java/ru/practicum/controller/category/CategoryController.java
@@ -1,4 +1,32 @@
 package ru.practicum.controller.category;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.controller.logs.ControllerLog;
+import ru.practicum.dto.category.CategoryDto;
+import ru.practicum.service.category.CategoryService;
+
+import java.util.List;
+
+@RestController
+@RequestMapping(path = "/categories")
+@RequiredArgsConstructor
 public class CategoryController {
+    private final CategoryService categoryService;
+
+    @ControllerLog
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public List<CategoryDto> getCategories(@RequestParam(value = "from", defaultValue = "0") Integer from,
+                                           @RequestParam(value = "size", defaultValue = "10") Integer size) {
+        return List.copyOf(categoryService.getAll(from, size));
+    }
+
+    @ControllerLog
+    @GetMapping("/{catId}")
+    @ResponseStatus(HttpStatus.OK)
+    public CategoryDto getCategoryById(@PathVariable("catId") Long catId) {
+        return categoryService.get(catId);
+    }
 }
